Get the list of validated packages from GitHub

This function downloads issues from github and extracts relevant information from them

get_valid_pkgs(approved_only = TRUE)

Arguments

approved_only

only return approved packages (logical)

Value

dataframe including package, version, etc.

Examples

# only approved packages
get_valid_pkgs()
#> [[1]]
#> [[1]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/16"
#> 
#> [[1]]$repository_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation"
#> 
#> [[1]]$labels_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/16/labels{/name}"
#> 
#> [[1]]$comments_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/16/comments"
#> 
#> [[1]]$events_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/16/events"
#> 
#> [[1]]$html_url
#> [1] "https://github.com/SwissClinicalTrialOrganisation/pkg_validation/issues/16"
#> 
#> [[1]]$id
#> [1] 2205161540
#> 
#> [[1]]$node_id
#> [1] "I_kwDOKHVskc6DcBhE"
#> 
#> [[1]]$number
#> [1] 16
#> 
#> [[1]]$title
#> [1] "[Package validation]: lme4 version 1.1-35.1"
#> 
#> [[1]]$user
#> [[1]]$user$login
#> [1] "aghaynes"
#> 
#> [[1]]$user$id
#> [1] 6594028
#> 
#> [[1]]$user$node_id
#> [1] "MDQ6VXNlcjY1OTQwMjg="
#> 
#> [[1]]$user$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/6594028?v=4"
#> 
#> [[1]]$user$gravatar_id
#> [1] ""
#> 
#> [[1]]$user$url
#> [1] "https://api.github.com/users/aghaynes"
#> 
#> [[1]]$user$html_url
#> [1] "https://github.com/aghaynes"
#> 
#> [[1]]$user$followers_url
#> [1] "https://api.github.com/users/aghaynes/followers"
#> 
#> [[1]]$user$following_url
#> [1] "https://api.github.com/users/aghaynes/following{/other_user}"
#> 
#> [[1]]$user$gists_url
#> [1] "https://api.github.com/users/aghaynes/gists{/gist_id}"
#> 
#> [[1]]$user$starred_url
#> [1] "https://api.github.com/users/aghaynes/starred{/owner}{/repo}"
#> 
#> [[1]]$user$subscriptions_url
#> [1] "https://api.github.com/users/aghaynes/subscriptions"
#> 
#> [[1]]$user$organizations_url
#> [1] "https://api.github.com/users/aghaynes/orgs"
#> 
#> [[1]]$user$repos_url
#> [1] "https://api.github.com/users/aghaynes/repos"
#> 
#> [[1]]$user$events_url
#> [1] "https://api.github.com/users/aghaynes/events{/privacy}"
#> 
#> [[1]]$user$received_events_url
#> [1] "https://api.github.com/users/aghaynes/received_events"
#> 
#> [[1]]$user$type
#> [1] "User"
#> 
#> [[1]]$user$site_admin
#> [1] FALSE
#> 
#> 
#> [[1]]$labels
#> [[1]]$labels[[1]]
#> [[1]]$labels[[1]]$id
#> [1] 5847585136
#> 
#> [[1]]$labels[[1]]$node_id
#> [1] "LA_kwDOKHVskc8AAAABXIsRcA"
#> 
#> [[1]]$labels[[1]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/labels/package"
#> 
#> [[1]]$labels[[1]]$name
#> [1] "package"
#> 
#> [[1]]$labels[[1]]$color
#> [1] "b60205"
#> 
#> [[1]]$labels[[1]]$default
#> [1] FALSE
#> 
#> [[1]]$labels[[1]]$description
#> [1] ""
#> 
#> 
#> [[1]]$labels[[2]]
#> [[1]]$labels[[2]]$id
#> [1] 5847585907
#> 
#> [[1]]$labels[[2]]$node_id
#> [1] "LA_kwDOKHVskc8AAAABXIsUcw"
#> 
#> [[1]]$labels[[2]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/labels/approved"
#> 
#> [[1]]$labels[[2]]$name
#> [1] "approved"
#> 
#> [[1]]$labels[[2]]$color
#> [1] "F4E375"
#> 
#> [[1]]$labels[[2]]$default
#> [1] FALSE
#> 
#> [[1]]$labels[[2]]$description
#> [1] ""
#> 
#> 
#> 
#> [[1]]$state
#> [1] "open"
#> 
#> [[1]]$locked
#> [1] FALSE
#> 
#> [[1]]$assignee
#> NULL
#> 
#> [[1]]$assignees
#> list()
#> 
#> [[1]]$milestone
#> NULL
#> 
#> [[1]]$comments
#> [1] 1
#> 
#> [[1]]$created_at
#> [1] "2024-03-25T08:21:05Z"
#> 
#> [[1]]$updated_at
#> [1] "2024-03-26T07:22:15Z"
#> 
#> [[1]]$closed_at
#> NULL
#> 
#> [[1]]$author_association
#> [1] "MEMBER"
#> 
#> [[1]]$active_lock_reason
#> NULL
#> 
#> [[1]]$body
#> [1] "### Name\r\n\r\nMC\r\n\r\n### What is your relationship with the SCTO?\r\n\r\nSCTO Stats and Methodology platform\r\n\r\n### Name of the package you have validated\r\n\r\nlme4\r\n\r\n### Version number of the package evaluated\r\n\r\n1.1-35.1\r\n\r\n### Date of release of the evaluated version of the package\r\n\r\n2023-11-05\r\n\r\n### The package author has...\r\n\r\nwell-known or known credentials\r\n\r\n### Is there a maintainer listed for the package and are their contact details available?\r\n\r\nAvailable\r\n\r\n### Package purpose\r\n\r\nStatistical; published\r\n\r\n### Number of dependencies\r\n\r\n11\r\n\r\n### Is the package on available from CRAN or bioconductor?\r\n\r\nYes\r\n\r\n### Is source code available, accessible and documented (i.e., well-structured and including comments) or is the source code unavailable or not clearly commented.\r\n\r\nYes\r\n\r\n### Number of downloads in the last 12 months\r\n\r\n5184335\r\n\r\n### Bug reporting address is available (in DESCRIPTION)\r\n\r\nYes\r\n\r\n### Does the package have one or more vignettes?\r\n\r\nYes\r\n\r\n### Does the package have unit and/or function tests performed by the authors? Are they comprehensive? Are they well documented?\r\n\r\nYes, comprehensive"
#> 
#> [[1]]$reactions
#> [[1]]$reactions$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/16/reactions"
#> 
#> [[1]]$reactions$total_count
#> [1] 0
#> 
#> [[1]]$reactions$`+1`
#> [1] 0
#> 
#> [[1]]$reactions$`-1`
#> [1] 0
#> 
#> [[1]]$reactions$laugh
#> [1] 0
#> 
#> [[1]]$reactions$hooray
#> [1] 0
#> 
#> [[1]]$reactions$confused
#> [1] 0
#> 
#> [[1]]$reactions$heart
#> [1] 0
#> 
#> [[1]]$reactions$rocket
#> [1] 0
#> 
#> [[1]]$reactions$eyes
#> [1] 0
#> 
#> 
#> [[1]]$timeline_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/16/timeline"
#> 
#> [[1]]$performed_via_github_app
#> NULL
#> 
#> [[1]]$state_reason
#> NULL
#> 
#> 
# all packages, regardless of status
get_valid_pkgs(FALSE)
#> Warning: NAs introduced by coercion
#> [[1]]
#> [[1]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/20"
#> 
#> [[1]]$repository_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation"
#> 
#> [[1]]$labels_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/20/labels{/name}"
#> 
#> [[1]]$comments_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/20/comments"
#> 
#> [[1]]$events_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/20/events"
#> 
#> [[1]]$html_url
#> [1] "https://github.com/SwissClinicalTrialOrganisation/pkg_validation/issues/20"
#> 
#> [[1]]$id
#> [1] 2207648616
#> 
#> [[1]]$node_id
#> [1] "I_kwDOKHVskc6Dlgto"
#> 
#> [[1]]$number
#> [1] 20
#> 
#> [[1]]$title
#> [1] "[Package validation]: dplyr version 1.1.4"
#> 
#> [[1]]$user
#> [[1]]$user$login
#> [1] "aghaynes"
#> 
#> [[1]]$user$id
#> [1] 6594028
#> 
#> [[1]]$user$node_id
#> [1] "MDQ6VXNlcjY1OTQwMjg="
#> 
#> [[1]]$user$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/6594028?v=4"
#> 
#> [[1]]$user$gravatar_id
#> [1] ""
#> 
#> [[1]]$user$url
#> [1] "https://api.github.com/users/aghaynes"
#> 
#> [[1]]$user$html_url
#> [1] "https://github.com/aghaynes"
#> 
#> [[1]]$user$followers_url
#> [1] "https://api.github.com/users/aghaynes/followers"
#> 
#> [[1]]$user$following_url
#> [1] "https://api.github.com/users/aghaynes/following{/other_user}"
#> 
#> [[1]]$user$gists_url
#> [1] "https://api.github.com/users/aghaynes/gists{/gist_id}"
#> 
#> [[1]]$user$starred_url
#> [1] "https://api.github.com/users/aghaynes/starred{/owner}{/repo}"
#> 
#> [[1]]$user$subscriptions_url
#> [1] "https://api.github.com/users/aghaynes/subscriptions"
#> 
#> [[1]]$user$organizations_url
#> [1] "https://api.github.com/users/aghaynes/orgs"
#> 
#> [[1]]$user$repos_url
#> [1] "https://api.github.com/users/aghaynes/repos"
#> 
#> [[1]]$user$events_url
#> [1] "https://api.github.com/users/aghaynes/events{/privacy}"
#> 
#> [[1]]$user$received_events_url
#> [1] "https://api.github.com/users/aghaynes/received_events"
#> 
#> [[1]]$user$type
#> [1] "User"
#> 
#> [[1]]$user$site_admin
#> [1] FALSE
#> 
#> 
#> [[1]]$labels
#> [[1]]$labels[[1]]
#> [[1]]$labels[[1]]$id
#> [1] 5847578368
#> 
#> [[1]]$labels[[1]]$node_id
#> [1] "LA_kwDOKHVskc8AAAABXIr3AA"
#> 
#> [[1]]$labels[[1]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/labels/triage"
#> 
#> [[1]]$labels[[1]]$name
#> [1] "triage"
#> 
#> [[1]]$labels[[1]]$color
#> [1] "61255E"
#> 
#> [[1]]$labels[[1]]$default
#> [1] FALSE
#> 
#> [[1]]$labels[[1]]$description
#> [1] ""
#> 
#> 
#> [[1]]$labels[[2]]
#> [[1]]$labels[[2]]$id
#> [1] 5847585136
#> 
#> [[1]]$labels[[2]]$node_id
#> [1] "LA_kwDOKHVskc8AAAABXIsRcA"
#> 
#> [[1]]$labels[[2]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/labels/package"
#> 
#> [[1]]$labels[[2]]$name
#> [1] "package"
#> 
#> [[1]]$labels[[2]]$color
#> [1] "b60205"
#> 
#> [[1]]$labels[[2]]$default
#> [1] FALSE
#> 
#> [[1]]$labels[[2]]$description
#> [1] ""
#> 
#> 
#> 
#> [[1]]$state
#> [1] "open"
#> 
#> [[1]]$locked
#> [1] FALSE
#> 
#> [[1]]$assignee
#> NULL
#> 
#> [[1]]$assignees
#> list()
#> 
#> [[1]]$milestone
#> NULL
#> 
#> [[1]]$comments
#> [1] 1
#> 
#> [[1]]$created_at
#> [1] "2024-03-26T09:07:34Z"
#> 
#> [[1]]$updated_at
#> [1] "2024-04-04T08:25:23Z"
#> 
#> [[1]]$closed_at
#> NULL
#> 
#> [[1]]$author_association
#> [1] "MEMBER"
#> 
#> [[1]]$active_lock_reason
#> NULL
#> 
#> [[1]]$body
#> [1] "### Name\r\n\r\nAlan Haynes\r\n\r\n### What is your relationship with the SCTO?\r\n\r\nSCTO Stats and Methodology platform\r\n\r\n### Name of the package you have validated\r\n\r\ndplyr\r\n\r\n### Version number of the package evaluated\r\n\r\n1.1.4\r\n\r\n### Date of release of the evaluated version of the package\r\n\r\n2023-11-17\r\n\r\n### The package author has...\r\n\r\nwell-known or known credentials\r\n\r\n### Is there a maintainer listed for the package and are their contact details available?\r\n\r\nAvailable\r\n\r\n### Package purpose\r\n\r\nNon-statistical\r\n\r\n### Number of dependencies\r\n\r\n13\r\n\r\n### Is the package on available from CRAN or bioconductor?\r\n\r\nYes\r\n\r\n### Is source code available, accessible and documented (i.e., well-structured and including comments) or is the source code unavailable or not clearly commented.\r\n\r\nYes\r\n\r\n### Number of downloads in the last 12 months\r\n\r\n18987564\r\n\r\n### Bug reporting address is available (in DESCRIPTION)\r\n\r\nYes\r\n\r\n### Does the package have one or more vignettes?\r\n\r\nYes\r\n\r\n### Does the package have unit and/or function tests performed by the authors? Are they comprehensive? Are they well documented?\r\n\r\nYes, comprehensive"
#> 
#> [[1]]$reactions
#> [[1]]$reactions$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/20/reactions"
#> 
#> [[1]]$reactions$total_count
#> [1] 0
#> 
#> [[1]]$reactions$`+1`
#> [1] 0
#> 
#> [[1]]$reactions$`-1`
#> [1] 0
#> 
#> [[1]]$reactions$laugh
#> [1] 0
#> 
#> [[1]]$reactions$hooray
#> [1] 0
#> 
#> [[1]]$reactions$confused
#> [1] 0
#> 
#> [[1]]$reactions$heart
#> [1] 0
#> 
#> [[1]]$reactions$rocket
#> [1] 0
#> 
#> [[1]]$reactions$eyes
#> [1] 0
#> 
#> 
#> [[1]]$timeline_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/20/timeline"
#> 
#> [[1]]$performed_via_github_app
#> NULL
#> 
#> [[1]]$state_reason
#> NULL
#> 
#> 
#> [[2]]
#> [[2]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/19"
#> 
#> [[2]]$repository_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation"
#> 
#> [[2]]$labels_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/19/labels{/name}"
#> 
#> [[2]]$comments_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/19/comments"
#> 
#> [[2]]$events_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/19/events"
#> 
#> [[2]]$html_url
#> [1] "https://github.com/SwissClinicalTrialOrganisation/pkg_validation/issues/19"
#> 
#> [[2]]$id
#> [1] 2205912385
#> 
#> [[2]]$node_id
#> [1] "I_kwDOKHVskc6De41B"
#> 
#> [[2]]$number
#> [1] 19
#> 
#> [[2]]$title
#> [1] "[Package validation]: TEST version 1"
#> 
#> [[2]]$user
#> [[2]]$user$login
#> [1] "aghaynes"
#> 
#> [[2]]$user$id
#> [1] 6594028
#> 
#> [[2]]$user$node_id
#> [1] "MDQ6VXNlcjY1OTQwMjg="
#> 
#> [[2]]$user$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/6594028?v=4"
#> 
#> [[2]]$user$gravatar_id
#> [1] ""
#> 
#> [[2]]$user$url
#> [1] "https://api.github.com/users/aghaynes"
#> 
#> [[2]]$user$html_url
#> [1] "https://github.com/aghaynes"
#> 
#> [[2]]$user$followers_url
#> [1] "https://api.github.com/users/aghaynes/followers"
#> 
#> [[2]]$user$following_url
#> [1] "https://api.github.com/users/aghaynes/following{/other_user}"
#> 
#> [[2]]$user$gists_url
#> [1] "https://api.github.com/users/aghaynes/gists{/gist_id}"
#> 
#> [[2]]$user$starred_url
#> [1] "https://api.github.com/users/aghaynes/starred{/owner}{/repo}"
#> 
#> [[2]]$user$subscriptions_url
#> [1] "https://api.github.com/users/aghaynes/subscriptions"
#> 
#> [[2]]$user$organizations_url
#> [1] "https://api.github.com/users/aghaynes/orgs"
#> 
#> [[2]]$user$repos_url
#> [1] "https://api.github.com/users/aghaynes/repos"
#> 
#> [[2]]$user$events_url
#> [1] "https://api.github.com/users/aghaynes/events{/privacy}"
#> 
#> [[2]]$user$received_events_url
#> [1] "https://api.github.com/users/aghaynes/received_events"
#> 
#> [[2]]$user$type
#> [1] "User"
#> 
#> [[2]]$user$site_admin
#> [1] FALSE
#> 
#> 
#> [[2]]$labels
#> [[2]]$labels[[1]]
#> [[2]]$labels[[1]]$id
#> [1] 5847374414
#> 
#> [[2]]$labels[[1]]$node_id
#> [1] "LA_kwDOKHVskc8AAAABXIfaTg"
#> 
#> [[2]]$labels[[1]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/labels/invalid"
#> 
#> [[2]]$labels[[1]]$name
#> [1] "invalid"
#> 
#> [[2]]$labels[[1]]$color
#> [1] "e4e669"
#> 
#> [[2]]$labels[[1]]$default
#> [1] TRUE
#> 
#> [[2]]$labels[[1]]$description
#> [1] "This doesn't seem right"
#> 
#> 
#> [[2]]$labels[[2]]
#> [[2]]$labels[[2]]$id
#> [1] 5847585136
#> 
#> [[2]]$labels[[2]]$node_id
#> [1] "LA_kwDOKHVskc8AAAABXIsRcA"
#> 
#> [[2]]$labels[[2]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/labels/package"
#> 
#> [[2]]$labels[[2]]$name
#> [1] "package"
#> 
#> [[2]]$labels[[2]]$color
#> [1] "b60205"
#> 
#> [[2]]$labels[[2]]$default
#> [1] FALSE
#> 
#> [[2]]$labels[[2]]$description
#> [1] ""
#> 
#> 
#> 
#> [[2]]$state
#> [1] "open"
#> 
#> [[2]]$locked
#> [1] FALSE
#> 
#> [[2]]$assignee
#> NULL
#> 
#> [[2]]$assignees
#> list()
#> 
#> [[2]]$milestone
#> NULL
#> 
#> [[2]]$comments
#> [1] 2
#> 
#> [[2]]$created_at
#> [1] "2024-03-25T14:36:00Z"
#> 
#> [[2]]$updated_at
#> [1] "2024-04-03T13:07:17Z"
#> 
#> [[2]]$closed_at
#> NULL
#> 
#> [[2]]$author_association
#> [1] "MEMBER"
#> 
#> [[2]]$active_lock_reason
#> NULL
#> 
#> [[2]]$body
#> [1] "### Name\r\n\r\nAH \r\n\r\n### What is your relationship with the SCTO?\r\n\r\nSCTO Stats and Methodology platform\r\n\r\n### Name of the package you have validated\r\n\r\nTEST\r\n\r\n### Version number of the package evaluated\r\n\r\n1\r\n\r\n### Date of release of the evaluated version of the package\r\n\r\n_No response_\r\n\r\n### The package author has...\r\n\r\ncredential\r\n\r\n### Is there a maintainer listed for the package and are their contact details available?\r\n\r\nAvailabl\r\n\r\n### Package purpose\r\n\r\nStatistical; publishe\r\n\r\n### Number of dependencies\r\n\r\n1\r\n\r\n### Is the package on available from CRAN or bioconductor?\r\n\r\nYe\r\n\r\n### Is source code available, accessible and documented (i.e., well-structured and including comments) or is the source code unavailable or not clearly commented.\r\n\r\nYe\r\n\r\n### Number of downloads in the last 12 months\r\n\r\n_No response_\r\n\r\n### Bug reporting address is available (in DESCRIPTION)\r\n\r\nYe\r\n\r\n### Does the package have one or more vignettes?\r\n\r\nYe\r\n\r\n### Does the package have unit and/or function tests performed by the authors? Are they comprehensive? Are they well documented?\r\n\r\nYes, comprehensiv"
#> 
#> [[2]]$reactions
#> [[2]]$reactions$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/19/reactions"
#> 
#> [[2]]$reactions$total_count
#> [1] 0
#> 
#> [[2]]$reactions$`+1`
#> [1] 0
#> 
#> [[2]]$reactions$`-1`
#> [1] 0
#> 
#> [[2]]$reactions$laugh
#> [1] 0
#> 
#> [[2]]$reactions$hooray
#> [1] 0
#> 
#> [[2]]$reactions$confused
#> [1] 0
#> 
#> [[2]]$reactions$heart
#> [1] 0
#> 
#> [[2]]$reactions$rocket
#> [1] 0
#> 
#> [[2]]$reactions$eyes
#> [1] 0
#> 
#> 
#> [[2]]$timeline_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/19/timeline"
#> 
#> [[2]]$performed_via_github_app
#> NULL
#> 
#> [[2]]$state_reason
#> NULL
#> 
#> 
#> [[3]]
#> [[3]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/18"
#> 
#> [[3]]$repository_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation"
#> 
#> [[3]]$labels_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/18/labels{/name}"
#> 
#> [[3]]$comments_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/18/comments"
#> 
#> [[3]]$events_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/18/events"
#> 
#> [[3]]$html_url
#> [1] "https://github.com/SwissClinicalTrialOrganisation/pkg_validation/issues/18"
#> 
#> [[3]]$id
#> [1] 2205169587
#> 
#> [[3]]$node_id
#> [1] "I_kwDOKHVskc6DcDez"
#> 
#> [[3]]$number
#> [1] 18
#> 
#> [[3]]$title
#> [1] "[Package validation]: tables version 0.9.17"
#> 
#> [[3]]$user
#> [[3]]$user$login
#> [1] "aghaynes"
#> 
#> [[3]]$user$id
#> [1] 6594028
#> 
#> [[3]]$user$node_id
#> [1] "MDQ6VXNlcjY1OTQwMjg="
#> 
#> [[3]]$user$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/6594028?v=4"
#> 
#> [[3]]$user$gravatar_id
#> [1] ""
#> 
#> [[3]]$user$url
#> [1] "https://api.github.com/users/aghaynes"
#> 
#> [[3]]$user$html_url
#> [1] "https://github.com/aghaynes"
#> 
#> [[3]]$user$followers_url
#> [1] "https://api.github.com/users/aghaynes/followers"
#> 
#> [[3]]$user$following_url
#> [1] "https://api.github.com/users/aghaynes/following{/other_user}"
#> 
#> [[3]]$user$gists_url
#> [1] "https://api.github.com/users/aghaynes/gists{/gist_id}"
#> 
#> [[3]]$user$starred_url
#> [1] "https://api.github.com/users/aghaynes/starred{/owner}{/repo}"
#> 
#> [[3]]$user$subscriptions_url
#> [1] "https://api.github.com/users/aghaynes/subscriptions"
#> 
#> [[3]]$user$organizations_url
#> [1] "https://api.github.com/users/aghaynes/orgs"
#> 
#> [[3]]$user$repos_url
#> [1] "https://api.github.com/users/aghaynes/repos"
#> 
#> [[3]]$user$events_url
#> [1] "https://api.github.com/users/aghaynes/events{/privacy}"
#> 
#> [[3]]$user$received_events_url
#> [1] "https://api.github.com/users/aghaynes/received_events"
#> 
#> [[3]]$user$type
#> [1] "User"
#> 
#> [[3]]$user$site_admin
#> [1] FALSE
#> 
#> 
#> [[3]]$labels
#> [[3]]$labels[[1]]
#> [[3]]$labels[[1]]$id
#> [1] 5847578368
#> 
#> [[3]]$labels[[1]]$node_id
#> [1] "LA_kwDOKHVskc8AAAABXIr3AA"
#> 
#> [[3]]$labels[[1]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/labels/triage"
#> 
#> [[3]]$labels[[1]]$name
#> [1] "triage"
#> 
#> [[3]]$labels[[1]]$color
#> [1] "61255E"
#> 
#> [[3]]$labels[[1]]$default
#> [1] FALSE
#> 
#> [[3]]$labels[[1]]$description
#> [1] ""
#> 
#> 
#> [[3]]$labels[[2]]
#> [[3]]$labels[[2]]$id
#> [1] 5847585136
#> 
#> [[3]]$labels[[2]]$node_id
#> [1] "LA_kwDOKHVskc8AAAABXIsRcA"
#> 
#> [[3]]$labels[[2]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/labels/package"
#> 
#> [[3]]$labels[[2]]$name
#> [1] "package"
#> 
#> [[3]]$labels[[2]]$color
#> [1] "b60205"
#> 
#> [[3]]$labels[[2]]$default
#> [1] FALSE
#> 
#> [[3]]$labels[[2]]$description
#> [1] ""
#> 
#> 
#> 
#> [[3]]$state
#> [1] "open"
#> 
#> [[3]]$locked
#> [1] FALSE
#> 
#> [[3]]$assignee
#> NULL
#> 
#> [[3]]$assignees
#> list()
#> 
#> [[3]]$milestone
#> NULL
#> 
#> [[3]]$comments
#> [1] 0
#> 
#> [[3]]$created_at
#> [1] "2024-03-25T08:26:06Z"
#> 
#> [[3]]$updated_at
#> [1] "2024-03-26T07:20:55Z"
#> 
#> [[3]]$closed_at
#> NULL
#> 
#> [[3]]$author_association
#> [1] "MEMBER"
#> 
#> [[3]]$active_lock_reason
#> NULL
#> 
#> [[3]]$body
#> [1] "### Name\r\n\r\nMC\r\n\r\n### What is your relationship with the SCTO?\r\n\r\nSCTO Stats and Methodology platform\r\n\r\n### Name of the package you have validated\r\n\r\ntables\r\n\r\n### Version number of the package evaluated\r\n\r\n0.9.17\r\n\r\n### Date of release of the evaluated version of the package\r\n\r\n2023-05-02\r\n\r\n### The package author has...\r\n\r\ncredentials\r\n\r\n### Is there a maintainer listed for the package and are their contact details available?\r\n\r\nAvailable\r\n\r\n### Package purpose\r\n\r\nStatistical; published\r\n\r\n### Number of dependencies\r\n\r\n4\r\n\r\n### Is the package on available from CRAN or bioconductor?\r\n\r\nYes\r\n\r\n### Is source code available, accessible and documented (i.e., well-structured and including comments) or is the source code unavailable or not clearly commented.\r\n\r\nYes\r\n\r\n### Number of downloads in the last 12 months\r\n\r\n148591\r\n\r\n### Bug reporting address is available (in DESCRIPTION)\r\n\r\nYes\r\n\r\n### Does the package have one or more vignettes?\r\n\r\nYes\r\n\r\n### Does the package have unit and/or function tests performed by the authors? Are they comprehensive? Are they well documented?\r\n\r\nNo"
#> 
#> [[3]]$reactions
#> [[3]]$reactions$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/18/reactions"
#> 
#> [[3]]$reactions$total_count
#> [1] 0
#> 
#> [[3]]$reactions$`+1`
#> [1] 0
#> 
#> [[3]]$reactions$`-1`
#> [1] 0
#> 
#> [[3]]$reactions$laugh
#> [1] 0
#> 
#> [[3]]$reactions$hooray
#> [1] 0
#> 
#> [[3]]$reactions$confused
#> [1] 0
#> 
#> [[3]]$reactions$heart
#> [1] 0
#> 
#> [[3]]$reactions$rocket
#> [1] 0
#> 
#> [[3]]$reactions$eyes
#> [1] 0
#> 
#> 
#> [[3]]$timeline_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/18/timeline"
#> 
#> [[3]]$performed_via_github_app
#> NULL
#> 
#> [[3]]$state_reason
#> NULL
#> 
#> 
#> [[4]]
#> [[4]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/17"
#> 
#> [[4]]$repository_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation"
#> 
#> [[4]]$labels_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/17/labels{/name}"
#> 
#> [[4]]$comments_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/17/comments"
#> 
#> [[4]]$events_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/17/events"
#> 
#> [[4]]$html_url
#> [1] "https://github.com/SwissClinicalTrialOrganisation/pkg_validation/issues/17"
#> 
#> [[4]]$id
#> [1] 2205165800
#> 
#> [[4]]$node_id
#> [1] "I_kwDOKHVskc6DcCjo"
#> 
#> [[4]]$number
#> [1] 17
#> 
#> [[4]]$title
#> [1] "[Package validation]: presize version 0.3.7"
#> 
#> [[4]]$user
#> [[4]]$user$login
#> [1] "aghaynes"
#> 
#> [[4]]$user$id
#> [1] 6594028
#> 
#> [[4]]$user$node_id
#> [1] "MDQ6VXNlcjY1OTQwMjg="
#> 
#> [[4]]$user$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/6594028?v=4"
#> 
#> [[4]]$user$gravatar_id
#> [1] ""
#> 
#> [[4]]$user$url
#> [1] "https://api.github.com/users/aghaynes"
#> 
#> [[4]]$user$html_url
#> [1] "https://github.com/aghaynes"
#> 
#> [[4]]$user$followers_url
#> [1] "https://api.github.com/users/aghaynes/followers"
#> 
#> [[4]]$user$following_url
#> [1] "https://api.github.com/users/aghaynes/following{/other_user}"
#> 
#> [[4]]$user$gists_url
#> [1] "https://api.github.com/users/aghaynes/gists{/gist_id}"
#> 
#> [[4]]$user$starred_url
#> [1] "https://api.github.com/users/aghaynes/starred{/owner}{/repo}"
#> 
#> [[4]]$user$subscriptions_url
#> [1] "https://api.github.com/users/aghaynes/subscriptions"
#> 
#> [[4]]$user$organizations_url
#> [1] "https://api.github.com/users/aghaynes/orgs"
#> 
#> [[4]]$user$repos_url
#> [1] "https://api.github.com/users/aghaynes/repos"
#> 
#> [[4]]$user$events_url
#> [1] "https://api.github.com/users/aghaynes/events{/privacy}"
#> 
#> [[4]]$user$received_events_url
#> [1] "https://api.github.com/users/aghaynes/received_events"
#> 
#> [[4]]$user$type
#> [1] "User"
#> 
#> [[4]]$user$site_admin
#> [1] FALSE
#> 
#> 
#> [[4]]$labels
#> [[4]]$labels[[1]]
#> [[4]]$labels[[1]]$id
#> [1] 5847578368
#> 
#> [[4]]$labels[[1]]$node_id
#> [1] "LA_kwDOKHVskc8AAAABXIr3AA"
#> 
#> [[4]]$labels[[1]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/labels/triage"
#> 
#> [[4]]$labels[[1]]$name
#> [1] "triage"
#> 
#> [[4]]$labels[[1]]$color
#> [1] "61255E"
#> 
#> [[4]]$labels[[1]]$default
#> [1] FALSE
#> 
#> [[4]]$labels[[1]]$description
#> [1] ""
#> 
#> 
#> [[4]]$labels[[2]]
#> [[4]]$labels[[2]]$id
#> [1] 5847585136
#> 
#> [[4]]$labels[[2]]$node_id
#> [1] "LA_kwDOKHVskc8AAAABXIsRcA"
#> 
#> [[4]]$labels[[2]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/labels/package"
#> 
#> [[4]]$labels[[2]]$name
#> [1] "package"
#> 
#> [[4]]$labels[[2]]$color
#> [1] "b60205"
#> 
#> [[4]]$labels[[2]]$default
#> [1] FALSE
#> 
#> [[4]]$labels[[2]]$description
#> [1] ""
#> 
#> 
#> 
#> [[4]]$state
#> [1] "open"
#> 
#> [[4]]$locked
#> [1] FALSE
#> 
#> [[4]]$assignee
#> NULL
#> 
#> [[4]]$assignees
#> list()
#> 
#> [[4]]$milestone
#> NULL
#> 
#> [[4]]$comments
#> [1] 0
#> 
#> [[4]]$created_at
#> [1] "2024-03-25T08:23:41Z"
#> 
#> [[4]]$updated_at
#> [1] "2024-03-25T10:37:59Z"
#> 
#> [[4]]$closed_at
#> NULL
#> 
#> [[4]]$author_association
#> [1] "MEMBER"
#> 
#> [[4]]$active_lock_reason
#> NULL
#> 
#> [[4]]$body
#> [1] "### Name\r\n\r\nMC\r\n\r\n### What is your relationship with the SCTO?\r\n\r\nSCTO Stats and Methodology platform\r\n\r\n### Name of the package you have validated\r\n\r\npresize\r\n\r\n### Version number of the package evaluated\r\n\r\n0.3.7\r\n\r\n### Date of release of the evaluated version of the package\r\n\r\n2023-02-27\r\n\r\n### The package author has...\r\n\r\ncredentials\r\n\r\n### Is there a maintainer listed for the package and are their contact details available?\r\n\r\nAvailable\r\n\r\n### Package purpose\r\n\r\nStatistical; published\r\n\r\n### Number of dependencies\r\n\r\n2\r\n\r\n### Is the package on available from CRAN or bioconductor?\r\n\r\nYes\r\n\r\n### Is source code available, accessible and documented (i.e., well-structured and including comments) or is the source code unavailable or not clearly commented.\r\n\r\nYes\r\n\r\n### Number of downloads in the last 12 months\r\n\r\n5768\r\n\r\n### Bug reporting address is available (in DESCRIPTION)\r\n\r\nYes\r\n\r\n### Does the package have one or more vignettes?\r\n\r\nYes\r\n\r\n### Does the package have unit and/or function tests performed by the authors? Are they comprehensive? Are they well documented?\r\n\r\nYes, comprehensive"
#> 
#> [[4]]$reactions
#> [[4]]$reactions$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/17/reactions"
#> 
#> [[4]]$reactions$total_count
#> [1] 0
#> 
#> [[4]]$reactions$`+1`
#> [1] 0
#> 
#> [[4]]$reactions$`-1`
#> [1] 0
#> 
#> [[4]]$reactions$laugh
#> [1] 0
#> 
#> [[4]]$reactions$hooray
#> [1] 0
#> 
#> [[4]]$reactions$confused
#> [1] 0
#> 
#> [[4]]$reactions$heart
#> [1] 0
#> 
#> [[4]]$reactions$rocket
#> [1] 0
#> 
#> [[4]]$reactions$eyes
#> [1] 0
#> 
#> 
#> [[4]]$timeline_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/17/timeline"
#> 
#> [[4]]$performed_via_github_app
#> NULL
#> 
#> [[4]]$state_reason
#> NULL
#> 
#> 
#> [[5]]
#> [[5]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/16"
#> 
#> [[5]]$repository_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation"
#> 
#> [[5]]$labels_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/16/labels{/name}"
#> 
#> [[5]]$comments_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/16/comments"
#> 
#> [[5]]$events_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/16/events"
#> 
#> [[5]]$html_url
#> [1] "https://github.com/SwissClinicalTrialOrganisation/pkg_validation/issues/16"
#> 
#> [[5]]$id
#> [1] 2205161540
#> 
#> [[5]]$node_id
#> [1] "I_kwDOKHVskc6DcBhE"
#> 
#> [[5]]$number
#> [1] 16
#> 
#> [[5]]$title
#> [1] "[Package validation]: lme4 version 1.1-35.1"
#> 
#> [[5]]$user
#> [[5]]$user$login
#> [1] "aghaynes"
#> 
#> [[5]]$user$id
#> [1] 6594028
#> 
#> [[5]]$user$node_id
#> [1] "MDQ6VXNlcjY1OTQwMjg="
#> 
#> [[5]]$user$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/6594028?v=4"
#> 
#> [[5]]$user$gravatar_id
#> [1] ""
#> 
#> [[5]]$user$url
#> [1] "https://api.github.com/users/aghaynes"
#> 
#> [[5]]$user$html_url
#> [1] "https://github.com/aghaynes"
#> 
#> [[5]]$user$followers_url
#> [1] "https://api.github.com/users/aghaynes/followers"
#> 
#> [[5]]$user$following_url
#> [1] "https://api.github.com/users/aghaynes/following{/other_user}"
#> 
#> [[5]]$user$gists_url
#> [1] "https://api.github.com/users/aghaynes/gists{/gist_id}"
#> 
#> [[5]]$user$starred_url
#> [1] "https://api.github.com/users/aghaynes/starred{/owner}{/repo}"
#> 
#> [[5]]$user$subscriptions_url
#> [1] "https://api.github.com/users/aghaynes/subscriptions"
#> 
#> [[5]]$user$organizations_url
#> [1] "https://api.github.com/users/aghaynes/orgs"
#> 
#> [[5]]$user$repos_url
#> [1] "https://api.github.com/users/aghaynes/repos"
#> 
#> [[5]]$user$events_url
#> [1] "https://api.github.com/users/aghaynes/events{/privacy}"
#> 
#> [[5]]$user$received_events_url
#> [1] "https://api.github.com/users/aghaynes/received_events"
#> 
#> [[5]]$user$type
#> [1] "User"
#> 
#> [[5]]$user$site_admin
#> [1] FALSE
#> 
#> 
#> [[5]]$labels
#> [[5]]$labels[[1]]
#> [[5]]$labels[[1]]$id
#> [1] 5847585136
#> 
#> [[5]]$labels[[1]]$node_id
#> [1] "LA_kwDOKHVskc8AAAABXIsRcA"
#> 
#> [[5]]$labels[[1]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/labels/package"
#> 
#> [[5]]$labels[[1]]$name
#> [1] "package"
#> 
#> [[5]]$labels[[1]]$color
#> [1] "b60205"
#> 
#> [[5]]$labels[[1]]$default
#> [1] FALSE
#> 
#> [[5]]$labels[[1]]$description
#> [1] ""
#> 
#> 
#> [[5]]$labels[[2]]
#> [[5]]$labels[[2]]$id
#> [1] 5847585907
#> 
#> [[5]]$labels[[2]]$node_id
#> [1] "LA_kwDOKHVskc8AAAABXIsUcw"
#> 
#> [[5]]$labels[[2]]$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/labels/approved"
#> 
#> [[5]]$labels[[2]]$name
#> [1] "approved"
#> 
#> [[5]]$labels[[2]]$color
#> [1] "F4E375"
#> 
#> [[5]]$labels[[2]]$default
#> [1] FALSE
#> 
#> [[5]]$labels[[2]]$description
#> [1] ""
#> 
#> 
#> 
#> [[5]]$state
#> [1] "open"
#> 
#> [[5]]$locked
#> [1] FALSE
#> 
#> [[5]]$assignee
#> NULL
#> 
#> [[5]]$assignees
#> list()
#> 
#> [[5]]$milestone
#> NULL
#> 
#> [[5]]$comments
#> [1] 1
#> 
#> [[5]]$created_at
#> [1] "2024-03-25T08:21:05Z"
#> 
#> [[5]]$updated_at
#> [1] "2024-03-26T07:22:15Z"
#> 
#> [[5]]$closed_at
#> NULL
#> 
#> [[5]]$author_association
#> [1] "MEMBER"
#> 
#> [[5]]$active_lock_reason
#> NULL
#> 
#> [[5]]$body
#> [1] "### Name\r\n\r\nMC\r\n\r\n### What is your relationship with the SCTO?\r\n\r\nSCTO Stats and Methodology platform\r\n\r\n### Name of the package you have validated\r\n\r\nlme4\r\n\r\n### Version number of the package evaluated\r\n\r\n1.1-35.1\r\n\r\n### Date of release of the evaluated version of the package\r\n\r\n2023-11-05\r\n\r\n### The package author has...\r\n\r\nwell-known or known credentials\r\n\r\n### Is there a maintainer listed for the package and are their contact details available?\r\n\r\nAvailable\r\n\r\n### Package purpose\r\n\r\nStatistical; published\r\n\r\n### Number of dependencies\r\n\r\n11\r\n\r\n### Is the package on available from CRAN or bioconductor?\r\n\r\nYes\r\n\r\n### Is source code available, accessible and documented (i.e., well-structured and including comments) or is the source code unavailable or not clearly commented.\r\n\r\nYes\r\n\r\n### Number of downloads in the last 12 months\r\n\r\n5184335\r\n\r\n### Bug reporting address is available (in DESCRIPTION)\r\n\r\nYes\r\n\r\n### Does the package have one or more vignettes?\r\n\r\nYes\r\n\r\n### Does the package have unit and/or function tests performed by the authors? Are they comprehensive? Are they well documented?\r\n\r\nYes, comprehensive"
#> 
#> [[5]]$reactions
#> [[5]]$reactions$url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/16/reactions"
#> 
#> [[5]]$reactions$total_count
#> [1] 0
#> 
#> [[5]]$reactions$`+1`
#> [1] 0
#> 
#> [[5]]$reactions$`-1`
#> [1] 0
#> 
#> [[5]]$reactions$laugh
#> [1] 0
#> 
#> [[5]]$reactions$hooray
#> [1] 0
#> 
#> [[5]]$reactions$confused
#> [1] 0
#> 
#> [[5]]$reactions$heart
#> [1] 0
#> 
#> [[5]]$reactions$rocket
#> [1] 0
#> 
#> [[5]]$reactions$eyes
#> [1] 0
#> 
#> 
#> [[5]]$timeline_url
#> [1] "https://api.github.com/repos/SwissClinicalTrialOrganisation/pkg_validation/issues/16/timeline"
#> 
#> [[5]]$performed_via_github_app
#> NULL
#> 
#> [[5]]$state_reason
#> NULL
#> 
#>