Interact with the GitHub API to get or set issue information

These functions provide various ways to interact with github issues.

Getting issues:

get_issue downloads a single issue from a repository.

get_issues downloads all issues from a repository.

get_test_reports downloads all issues from a repository, filtering those with test labels, and converts the issue to a dataframe.

Labels:

add_label adds a particular label to an issue.

remove_label removes a particular label from an issue.

Comments:

get_comments downloads all comments from an issue.

post_comment posts a comment to an issue.

Closing an issue:

close_issue closes an issue.

These functions are primarily for the use in github actions rather than for the standard user.

add_label(issue, label, repo = sctoreports())

close_issue(issue, repo = sctoreports())

get_comments(issue, repo = sctoreports())

get_issue(issue, repo = sctoreports())

get_issues(repo = sctoreports())

get_test_reports(repo = sctoreports(), approved_only = FALSE)

post_comment(issue, comment, repo = sctoreports())

remove_label(issue, label, repo = sctoreports())

Arguments

issue

issue number

label

label to remove

repo

repository name

approved_only

only return approved packages (logical)

comment

the comment to be made

Value

a list of comments

dataframe

Examples

# remove_label(issue = 1, label = "test")

# post_comment(issue = 1, comment = "This is a test comment")
# get_issue(21)
# post_comment(issue = 1, comment = "This is a test comment")
# remove_label(issue = 1, label = "test")